adding tests
